+++ app/utils.go
@@ -1,4 +1,4 @@
-package cli
+package app
 
 import (
 	"errors"
