+++ app/utils_test.go
@@ -1,4 +1,4 @@
-package cli
+package app
 
 import (
 	"bytes"
